Copied the functional test stuff from the attendance branch.
