+++ .travis.yml
@@ -41,7 +41,7 @@ script:
     - python web2py.py --no-banner --nogui -a bob -i 127.0.0.1 -p 8000 &
     - sleep 3
     - cd applications/student_growth_tracker/
-    - python -m unittest test.teacher_functional_tests
+    - python -m unittest tests.teacher_functional_tests
     - python -m unittest tests.student_functional_tests
     - python -m unittest tests.parent_functional_tests
     - cd ~/web2py
